requirements.txt
flet
requests
pillow
pyperclip